fix(reports): reply polling gets its own per-key limiter — a class behind one NAT can poll and still report (#1352)
GET /api/replies and POST /api/replies/ack shared the per-IP ingest limiter
(10/min) with POST /api/bugreport, so 25 installs behind one school NAT polling
after world start ate the budget — and a real F1 report from that network in the
same minute bounced with 429 (spooled, re-sent later; delayed, not lost).

- Separate fixed-window limiter keyed by reply key (BBS_REPORTS_REPLY_PER_MINUTE,
  default 30) via GuardReplyKey; the per-IP limiter guards report submission only.
  Malformed keys are rejected before they can occupy a counter.
- Routes moved from Program.cs into ReportHostApp.Create(config, store, notifier,
  args) so the test suite can start the real app in-process on a loopback port;
  Program.cs only reads the environment and runs it.
- ReportHostHttpTests (4, real HTTP): 30 keys behind one IP all poll and a report
  still lands while the ingest limiter stays on; per-key not per-IP; poll → ack →
  answer end to end; write key still required.
- Docs: REPORT_HOST.md (config, code map), PLAYER_FEEDBACK.md, TODO.md, CHANGELOG.md.
  Needs a ReportHost image redeploy (reports-image.yml).
